requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
matplotlib==3.7.2
numpy==1.24.3
gunicorn==21.2.0
